pyproject.toml
[tool.poetry]
name = "compilex"
version = "0.1.0"
description = ""
authors = ["SkMishra <sanatkumarmishra02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
channels = {extras = ["daphne"], version = "^4.1.0"}
psutil = "^6.0.0"
aioboto3 = "^13.1.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
